requirements.txt
tensorflow>=2.10.0
streamlit>=1.28.0
opencv-python>=4.8.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
pillow>=10.0.0
plotly>=5.17.0
gradio>=3.50.0
streamlit-option-menu>=0.3.6
streamlit-image-comparison>=0.1.0